dtype for fracs output
Former-commit-id: 895bf6e8fd29c3540c79661cf1a2405c39c94850
Former-commit-id: d1b642b4b38e3e717353eb05f2098c64851c724e

+++ toblerone/core.py
@@ -807,7 +807,7 @@ def _estimateFractions(surf, supersampler, descriptor, cores):
         raise workerFractions[0]
 
     # Clip results to 1 (reqd due to geometric approximations)
-    fractions = np.concatenate(workerFractions, dtype=np.float32)
+    fractions = np.concatenate(workerFractions)
     return np.minimum(fractions, 1.0)
